Add files via upload
This is a ChatBot for Automation with basic functionalities I created a long time ago. It can tell you the time, change its voice, read documents, and open websites and Apps for you. 
It has some technical errors, which I am looking forward to removing in upcoming versions. Let me know if you can make any improvements to it.